+++ src/components/scorecard/components/match_types.js
@@ -3,6 +3,9 @@ exports.SHOOTOUT = 2;
 exports.X01HANDICAP = 3;
 exports.CRICKET = 4;
 exports.DARTS_AT_X = 5;
+exports.AROUND_THE_WORLD = 6;
+exports.SHANGHAI = 7;
+exports.AROUND_THE_CLOCK = 8;
 
 // Define types which support a simple input mode
-exports.SUPPORT_SIMPLE_INPUT = [ 5 ];
+exports.SUPPORT_SIMPLE_INPUT = [ 5, 6, 7, 8 ];