+++ src/components/scorecard/scorecard.component.js
@@ -1,8 +1,12 @@
 const alertify = require("../../util/alertify");
+
 var x01 = require("./components/x01");
 var shootout = require("./components/shootout");
 var cricket = require("./components/cricket");
 var dartsAtX = require("./components/darts_at_x");
+var aroundTheWorld = require("./components/around_the_world");
+var shanghai = require("./components/shanghai");
+var aroundTheClock = require("./components/around_the_clock");
 var types = require("./components/match_types");
 
 const DART_CONTAINER_MAP = { 1: 'first', 2: 'second', 3: 'third' };
@@ -11,6 +15,7 @@ module.exports = {
     onCreate(input) {
         var player = input.player;
         this.state = {
+            uuid: input.uuid,
             players: input.players,
             leg: input.leg,
             player: player,
@@ -71,7 +76,7 @@ module.exports = {
         return this.state.currentDart;
     },
 
-    removeLast() {
+    removeLast(external) {
         if (this.state.currentDart <= 1 && this.state.isSubmitted) {
             this.emit('undo-throw');
             return;
@@ -82,7 +87,7 @@ module.exports = {
 
             switch (this.state.type) {
                 case types.SHOOTOUT:
-                    shootout.removeLast.bind(this)(dart);
+                    shootout.removeLast.bind(this)(dart, external);
                     break;
                 case types.X01:
                     var value = dart.getValue();
@@ -92,10 +97,19 @@ module.exports = {
                     this.emit('possible-throw', false, false, this.state.currentDart, -dart.getScore(), dart.getMultiplier(), true);
                     break;
                 case types.CRICKET:
-                    cricket.removeLast.bind(this)(dart);
+                    cricket.removeLast.bind(this)(dart, external);
                     break;
                 case types.DARTS_AT_X:
-                    dartsAtX.removeLast.bind(this)(dart);
+                    dartsAtX.removeLast.bind(this)(dart, external);
+                    break;
+                case types.AROUND_THE_WORLD:
+                    aroundTheWorld.removeLast.bind(this)(dart, external);
+                    break;
+                case types.SHANGHAI:
+                    shanghai.removeLast.bind(this)(dart, external);
+                    break;
+                case types.AROUND_THE_CLOCK:
+                    aroundTheClock.removeLast.bind(this)(dart, external);
                     break;
             }
             dart.reset();
@@ -121,6 +135,15 @@ module.exports = {
                     break;
                 case types.DARTS_AT_X:
                     submitting = dartsAtX.confirmThrow.bind(this)(external);
+                    break;
+                case types.AROUND_THE_WORLD:
+                    submitting = aroundTheWorld.confirmThrow.bind(this)(external);
+                    break;
+                case types.SHANGHAI:
+                    submitting = shanghai.confirmThrow.bind(this)(external);
+                    break;
+                case types.AROUND_THE_CLOCK:
+                    submitting = aroundTheClock.confirmThrow.bind(this)(external);
                     break;
             }
         }