+++ src/pages/player/components/player/components/tab-statistics/tab-statistics.component.js
@@ -1,0 +1,5 @@
+module.exports = {
+    typeChanged(typeId) {
+        this.emit('type-changed', typeId);
+    }
+};