+++ src/pages/players/components/players-table/players-table.component.js
@@ -1,5 +1,4 @@
 module.exports = {
-
     viewPlayer(event, element) {
         var playerId = element.attributes['data-player'].value;
         location.href = '/players/' + playerId + '/statistics'